fix: refactor transaction item
