--- frontend/src/components/transaction-history/transaction-history-item.tsx
@@ -1,11 +1,0 @@
-import { backend as BackendModel } from '../../wailsjs/go/models'
-
-type TransactionHistoryItemProps = {
-  transaction: BackendModel.SentTransaction
-}
-
-export const TransactionHistoryItem = ({
-  transaction
-}: TransactionHistoryItemProps) => {
-  return <div>{JSON.stringify(transaction)}</div>
-}
