+++ frontend/src/components/transaction-item/index.ts
@@ -1,0 +1,1 @@
+export * from './transaction-item'
