--- frontend/src/components/transaction-manager/transaction-types.ts
@@ -1,35 +1,0 @@
-export enum TransactionKeys {
-  UNKNOWN = 'unknown',
-  ORDER_SUBMISSION = 'orderSubmission',
-  ORDER_CANCELLATION = 'orderCancellation',
-  ORDER_AMENDMENT = 'orderAmendment',
-  VOTE_SUBMISSION = 'voteSubmission',
-  WITHDRAW_SUBMISSION = 'withdrawSubmission',
-  LIQUIDTY_PROVISION_SUBMISSION = 'liquidityProvisionSubmission',
-  LIQUIDTY_PROVISION_CANCELLATION = 'liquidityProvisionCancellation',
-  LIQUIDITY_PROVISION_AMENDMENT = 'liquidityProvisionAmendment',
-  PROPOSAL_SUBMISSION = 'proposalSubmission',
-  ANNOUNCE_NODE = 'announceNode',
-  NODE_VOTE = 'nodeVote',
-  NODE_SIGNATURE = 'nodeSignature',
-  CHAIN_EVENT = 'chainEvent',
-  ORACLE_DATA_SUBMISSION = 'oracleDataSubmission',
-  UNDELEGATE_SUBMISSION = 'undelegateSubmission',
-  DELEGATE_SUBMISSION = 'delegateSubmission',
-  TRANSFER = 'transfer',
-  CANCEL_TRANSFER = 'cancelTransfer',
-  KEY_ROTATE_SUBMISSION = 'keyRotateSubmission',
-  ETHEREUM_KEY_ROTATE_SUBMISSION = 'ethereumKeyRotateSubmission'
-}
-
-export interface ParsedTx {
-  txId: string // signature
-  tx: object // payload
-  txHash: string | null
-  type: TransactionKeys
-  receivedAt: Date
-  pubKey: string
-  pending: boolean
-  sentAt: Date | null
-  error: string | null
-}
