--- frontend/src/components/transaction-queue/transaction-queue-item.tsx
@@ -1,11 +1,0 @@
-import { backend as BackendModel } from '../../wailsjs/go/models'
-
-type TransactionQueueItemProps = {
-  transaction: BackendModel.ConsentRequest
-}
-
-export const TransactionQueueItem = ({
-  transaction
-}: TransactionQueueItemProps) => {
-  return <div>{JSON.stringify(transaction)}</div>
-}
